requirements.txt
appdirs==1.4.4
attrs==19.3.0
aws-cdk.aws-events==1.38.0
aws-cdk.aws-iam==1.38.0
aws-cdk.aws-kms==1.38.0
aws-cdk.aws-s3==1.38.0
aws-cdk.cloud-assembly-schema==1.38.0
aws-cdk.core==1.38.0
aws-cdk.cx-api==1.38.0
aws-cdk.region-info==1.38.0
black==19.10b0
cattrs==1.0.0
click==7.1.2
constructs==3.0.3
# Editable Git install with no remote (cookies==0.0.1)
-e .
jsii==1.5.0
more-itertools==8.2.0
packaging==20.3
pathspec==0.8.0
pluggy==0.13.1
publication==0.0.3
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
regex==2020.5.7
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
wcwidth==0.1.9
